+++ apps/backend/src/routers/recipes.ts
@@ -1,7 +1,20 @@
 import { zValidator } from '@hono/zod-validator'
+import { z } from 'zod'
 import { honoFactory } from '../factory'
 import { getRecipes } from '../getRecipes'
-import { cookpadSearchParamSchema } from '../schemas/cookpadSearchParamSchema'
+
+const numericString = z.string().regex(/^\d+$/, { message: 'This field must be a number' })
+const cookpadSearchParamSchema = z.object({
+  ingredients: z
+    .union([z.array(z.string().min(1)).min(1), z.string().min(1)])
+    .transform((v) => (Array.isArray(v) ? v : [v]))
+    .describe('検索する食材'),
+  page: numericString
+    .optional()
+    .default('1')
+    .describe('ページ番号')
+    .transform((v) => Number.parseInt(v)),
+})
 
 const recipesRouter = honoFactory
   .createApp()
